+++ apps/zigbee/multisensor_deepSleep/firmware/src/app_idle_task.c
@@ -138,8 +138,11 @@ void app_idle_task( void )
               }
               else
               {
-                RF_Timer_Cal(WSS_ENABLE_ZB);
+            
+                   RF_Timer_Cal(WSS_ENABLE_ZB);
+            
               }
+
             }
         }
     }
@@ -199,7 +202,7 @@ static void app_idle_setRtcTimeout(TickType_t expectedIdleTick, uint32_t current
        2. RTC Clock : RTC_Timer32FrequencyGet
        3. expectedIdleTime (ms) * RTC clock (32 kHz) = compareValue value
     */
-    compareValue = (expectedIdleTick * RTC_Timer32FrequencyGet() + (configTICK_RATE_HZ / 2)) / configTICK_RATE_HZ;
+    compareValue = ((uint64_t)expectedIdleTick * RTC_Timer32FrequencyGet() + (configTICK_RATE_HZ / 2)) / configTICK_RATE_HZ;
 
 
     /* Give a compensation value to eliminate the offset between RTC and system timer
